+++ Public/Get-MarkdownCodeBlock.ps1
@@ -2,18 +2,43 @@ function Get-MarkdownCodeBlock {
     <#
         .SYNOPSIS
             Extracts PowerShell code blocks from markdown
+        
         .EXAMPLE
         Get-MarkdownCodeBlock -Path C:\temp\myfile.md
+        
+        .Example
+        $url = 'https://private.url.com/test.md'
+        $header = @{"Authorization"="token $($env:GITHUB_TOKEN)"}
+        Get-MarkdownCodeBlock $url $header        
     #>
     param(
         [Parameter(ValueFromPipelineByPropertyName, ValueFromPipeline)]
         [Alias('FullName')]
-        $Path
+        $Path,
+        $Headers
     )
 
     Process {
         if (([System.Uri]::IsWellFormedUriString($Path, 'Absolute'))) {
-            $mdContent = Invoke-RestMethod $Path
+            $InvokeParams = @{Uri = $Path }
+            
+            if ($Headers) {
+                $InvokeParams["Headers"] = $Headers
+            }
+
+            try {
+                $Error.Clear()
+                $mdContent = Invoke-RestMethod @InvokeParams
+            }
+            catch {
+                $err = [PSCustomObject]@{
+                    Path  = $Path
+                    Error = $_
+                }
+                
+                return $err
+            }
+            
             $mdContent = $mdContent -split "`n"
         }
         elseif (Test-Path $Path) {